backend/requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.0
Flask-Cors==5.0.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.31.0
webdriver-manager==4.0.2
gunicorn==23.0.0